Cleaned up warnings

+++ src/main/java/com/ignis/igrobotics/client/screen/elements/ScrollableElement.java
@@ -84,7 +84,7 @@ public class ScrollableElement extends GuiElement {
         if(Screen.hasShiftDown()) {
             int scroll_diff = scrollDistanceX;
 
-            scrollDistanceX += delta / -10;
+            scrollDistanceX += (int) (delta / -10);
             scrollDistanceX = Math.max(Math.min(scrollDistanceX, scrollMaxX - width), 0);
 
             scroll_diff = scroll_diff - scrollDistanceX;
@@ -99,7 +99,7 @@ public class ScrollableElement extends GuiElement {
         } else {
             int scroll_diff = scrollDistanceY;
 
-            scrollDistanceY += delta * -10;
+            scrollDistanceY += (int) (delta * -10);
             scrollDistanceY = Math.max(Math.min(scrollDistanceY, scrollMaxY - height), 0);
 
             scroll_diff = scroll_diff - scrollDistanceY;
